refactor(dart): Runtime using AOT
