+++ dart/Dockerfile
@@ -1,15 +1,22 @@
-FROM dart:3.2
+FROM dart:3.2 AS build
 
+# Set the working directory
 WORKDIR /app
 
-ADD pubspec.yaml /app/
-
+# Add pubspec.yaml and get dependencies
+ADD pubspec.* ./
 RUN dart pub get --no-precompile
 
-ADD . /app/
-
+# Copy app source code and AOT compile it
+ADD . .
 RUN dart pub get --offline --no-precompile
 
-WORKDIR /app
+# AOT compile `server.dart` to server executable
+RUN dart compile exe server.dart -o server
+
+# Build minimal serving image from AOT-compiled `server` executable
+FROM scratch
+COPY --from=build /runtime/ /
+COPY --from=build /app/server /app/server
 
 CMD {{command}}
